requirements.txt
biopython
chakin
cython
gym
googledrivedownloader
h5py
imageio
jieba
keras
lightdock
matplotlib
moviepy
mpi4py
networkx
nltk
numpy
plyfile
pybullet>=2.8.1
pygame
pymatgen==2020.4.29
pyqt5
pyqtgraph
pyro-ppl>=0.4.0
pytest>=5.0.0
pytorch-transformers>=1.1.0
ray[rllib,debug]>=1.0.0
rdflib
scipy
spacy>=2.1.6
tensorboard
torch-geometric
tqdm
